refactor: add feature-toggles as a module
