+++ src/index.tsx
@@ -38,7 +38,7 @@ import {PopOverContextProvider} from '@atb/popover';
 import {StorybookContextProvider} from '@atb/storybook/StorybookContext';
 import {NotificationContextProvider} from './notifications';
 import {BeaconsContextProvider} from './beacons/BeaconsContext';
-import {FeatureTogglesContextProvider} from '@atb/feature-toggles';
+import {FeatureTogglesContextProvider} from '@atb/modules/feature-toggles';
 import {configureReanimatedLogger} from 'react-native-reanimated';
 
 configureAndStartBugsnag();
