+++ src/loading-screen/LoadingScreenBoundary.tsx
@@ -4,7 +4,7 @@ import {LoadingErrorScreen} from './LoadingErrorScreen';
 import {useDelayGate} from '@atb/utils/use-delay-gate';
 import {useLoadingState} from '@atb/loading-screen/use-loading-state';
 import {useNotifyBugsnagOnTimeoutStatus} from '@atb/loading-screen/use-notify-bugsnag-on-timeout-status';
-import {useFeatureTogglesContext} from '@atb/feature-toggles';
+import {useFeatureTogglesContext} from '@atb/modules/feature-toggles';
 import {useRemoteConfigContext} from '@atb/RemoteConfigContext';
 
 const LOADING_TIMEOUT_MS = 10000;
