+++ src/mobility/queries/use-active-shmo-booking-query.tsx
@@ -2,7 +2,7 @@ import {useQuery} from '@tanstack/react-query';
 import {getActiveShmoBooking} from '@atb/api/mobility';
 import {ONE_MINUTE_MS} from '@atb/utils/durations';
 import {useAcceptLanguage} from '@atb/api/use-accept-language';
-import {useFeatureTogglesContext} from '@atb/feature-toggles';
+import {useFeatureTogglesContext} from '@atb/modules/feature-toggles';
 
 export const getActiveShmoBookingQueryKey = (acceptLanguage: string) => [
   'GET_ACTIVE_SHMO_BOOKING_QUERY_KEY',
