+++ src/modules/feature-toggles/__tests__/useFeatureTogglesContextState.test.ts
@@ -1,5 +1,5 @@
 import {act, renderHook} from '@testing-library/react-hooks';
-import {useFeatureTogglesContextState} from '../use-feature-toggle-context-state';
+import {useFeatureTogglesContextState} from '../use-feature-toggle-context-state.ts';
 import type {StorageService} from '@atb/storage';
 import type {RemoteConfig} from '@atb/remote-config';
 import {toggleSpecifications} from '../toggle-specifications.ts';
