+++ src/stacks-hierarchy/Root_TicketInformationScreen.tsx
@@ -16,7 +16,7 @@ import {TipsAndInformation} from '@atb/tips-and-information';
 import {useSafeAreaInsets} from 'react-native-safe-area-context';
 import {useOperatorBenefitsForFareProduct} from '@atb/mobility/use-operator-benefits-for-fare-product';
 import {MobilitySingleBenefitInfoSectionItem} from '@atb/mobility/components/MobilitySingleBenefitInfoSectionItem';
-import {useFeatureTogglesContext} from '@atb/feature-toggles';
+import {useFeatureTogglesContext} from '@atb/modules/feature-toggles';
 
 type Props = RootStackScreenProps<'Root_TicketInformationScreen'>;
