+++ src/travel-details-screens/components/Trip.tsx
@@ -47,9 +47,9 @@ import {ScreenReaderAnnouncement} from '@atb/components/screen-reader-announceme
 import {getAxiosErrorType} from '@atb/api/utils';
 import {FormFactor} from '@atb/api/types/generated/mobility-types_v2';
 import {isDefined} from '@atb/utils/presence';
-import {useFeatureTogglesContext} from '@atb/feature-toggles';
 import {useInAppReviewFlow} from '@atb/utils/use-in-app-review';
 import {useFocusEffect} from '@react-navigation/native';
+import {useFeatureTogglesContext} from '@atb/modules/feature-toggles';
 
 export type TripProps = {
   tripPattern: TripPattern;
