+++ src/utils/use-in-app-review.ts
@@ -1,6 +1,6 @@
 import InAppReview from 'react-native-in-app-review';
 import {useAnalyticsContext} from '@atb/analytics';
-import {useFeatureTogglesContext} from '@atb/feature-toggles';
+import {useFeatureTogglesContext} from '@atb/modules/feature-toggles';
 import {useCallback} from 'react';
 import {notifyBugsnag} from '@atb/utils/bugsnag-utils';
 import {storage} from '@atb/storage';
